chore(examples): regen catch-up after rebase onto v2.20
Maintainer follow-up to the prior regen commit. After this branch was
rebased onto v2.20 main, codegen output shifted under:

- S1.1 forall harness lowering (real `<prop>_at(s, binder)` predicates
  + `kani::any::<T>()` bindings replacing the legacy vacuous `true`
  stub) — affects percolator/tests/kani.rs + proptest.rs.
- v2.14+ body-less invariant change — escrow-split's `conservation`
  invariant now emits a documentation comment instead of the
  tautological `theorem conservation : True := trivial`.
- PR #45's own state_binder / type-aware seed-state / mul_div / tuple-
  arity / recursion_limit fixes combined with v2.20's match-block
  proptest path — affects lending/tests/proptest.rs,
  multisig/{tests,programs/tests}/proptest.rs, and
  percolator/programs/src/guards.rs.

`qedgen check --regen-drift` now reports "drift check clean —
5 example(s) checked". CI gates (fmt, clippy -- -D warnings,
version-consistency) clean.

+++ examples/rust/percolator/tests/kani.rs
@@ -71,7 +71,7 @@ fn account_solvent(_s: &State) -> bool {
 /// account_solvent: per-slot check at `i: AccountIdx` (v2.20 forall lowering)
 #[allow(unused_variables)]
 fn account_solvent_at(s: &State, i: usize) -> bool {
-    (!(s.accounts[(i) as usize].active == 1)) || (((s.accounts[(i) as usize].capital) as i128) + s.accounts[(i) as usize].pnl >= ((0) as i128))
+    (!(s.accounts[(i) as usize].active == 1)) || (((((s.accounts[(i) as usize].capital) as i128) + ((s.accounts[(i) as usize].pnl) as i128)) as i128) >= ((0) as i128))
 }
 
 // ============================================================================
